Rates and Ratios

Calculate attack rate, case fatality rate, and mortality rate

attack_rate(
  cases,
  population,
  conf_level = 0.95,
  multiplier = 100,
  mergeCI = FALSE,
  digits = 2
)

case_fatality_rate(
  deaths,
  population,
  conf_level = 0.95,
  multiplier = 100,
  mergeCI = FALSE,
  digits = 2
)

case_fatality_rate_df(
  x,
  deaths,
  group = NULL,
  conf_level = 0.95,
  multiplier = 100,
  mergeCI = FALSE,
  digits = 2,
  add_total = FALSE
)

mortality_rate(
  deaths,
  population,
  conf_level = 0.95,
  multiplier = 10^4,
  mergeCI = FALSE,
  digits = 2
)

Arguments

cases, deaths

number of cases or deaths in a population. For _df functions, this can be the name of a logical column OR an evaluated logical expression (see examples).

population

the number of individuals in the population.

conf_level

a number representing the confidence level for which to calculate the confidence interval. Defaults to 0.95, representing a 95% confidence interval using binom::binom.wilson()

multiplier

The base by which to multiply the output:

multiplier = 1: ratio between 0 and 1

multiplier = 100: proportion

multiplier = 10^4: x per 10,000 people

mergeCI

Whether or not to put the confidence intervals in one column (default is FALSE)

digits

if mergeCI = TRUE, this determines how many digits are printed

x

a data frame

group

the bare name of a column to use for stratifying the output

add_total

if group is not NULL, then this will add a row containing the total value across all groups.

Value

a data frame with five columns that represent the numerator, denominator, rate, lower bound, and upper bound.

attack_rate(): cases, population, ar, lower, upper

case_fatality_rate(): deaths, population, cfr, lower, upper

Examples

# Attack rates can be calculated with just two numbers
print(ar <- attack_rate(10, 50), digits = 4) # 20% attack rate
#>   cases population ar lower upper
#> 1    10         50 20 11.24 33.04

# print them inline using `fmt_ci_df()`
fmt_ci_df(ar)
#> Error in fmt_ci_df(ar): could not find function "fmt_ci_df"

# Alternatively, if you want one column for the CI, use `mergeCI = TRUE`
attack_rate(10, 50, mergeCI = TRUE, digits = 2) # 20% attack rate
#> Error in merge_ci_df(res, digits = digits): could not find function "merge_ci_df"

print(cfr <- case_fatality_rate(1, 100), digits = 2) # CFR of 1%
#>   deaths population cfr lower upper
#> 1      1        100   1  0.18   5.4
fmt_ci_df(cfr)
#> Error in fmt_ci_df(cfr): could not find function "fmt_ci_df"

# using a data frame
if (require("outbreaks")) {
  withAutoprint({
  e <- outbreaks::ebola_sim$linelist
  case_fatality_rate_df(e,
    outcome == "Death",
    group = gender,
    add_total = TRUE,
    mergeCI = TRUE
  )
  })
}
#> Loading required package: outbreaks
#> > e <- outbreaks::ebola_sim$linelist
#> > case_fatality_rate_df(e, outcome == "Death", group = gender, add_total = TRUE, 
#> +     mergeCI = TRUE)
#> Error in dplyr::summarise(x, `:=`(!!quote(deaths), sum(!!qdeath, na.rm = TRUE)),     `:=`(!!quote(population), dplyr::n() - sum(is.na(!!qdeath))),     `:=`(!!quote(cfr), list(case_fatality_rate(.data$deaths,         .data$population, conf_level, multiplier, mergeCI, digits)[-(1:2)]))): Problem while computing `cfr = list(...)`.
#> ℹ The error occurred in group 1: gender = f.
#> Caused by error in `merge_ci_df()`:
#> ! could not find function "merge_ci_df"