An attack rate wrapper function (using gtsummary and epikit packages)that takes a gtsummary object and returns a gtsummary object with attack rate (per given multiple) with 95% confidence interval

An attack rate wrapper function (using gtsummary and epikit packages)that takes a gtsummary object and returns a gtsummary object with attack rate (per given multiple) with 95% confidence interval

An attack rate wrapper function (using gtsummary and epikit packages)that takes a gtsummary object and returns a gtsummary object withattack rate (per given multiple) with 95% confidence interval

An case fatality rate wrapper function (using gtsummary and epikit packages) that takes a gtsummary object and returns a gtsummary object with number of deaths, case fatality rate, and 95% confidence interval.

A gtsummary wrapper function that takes a gtsummary object and removes a column from the table body by column name

A gtsummary wrapper function that takes a data frame and adds cross tabs by exposure and outcome

A function that adds mh odds ratio to an existing gtsummary object with same dimensions (will add to this later.)

A case fatality rate wrapper function to be passed to the gtsummary::add_stat function, which returns a data frame with a single row to be used with dichotomous data or overall data.Calls epikit::case_fatality_rate_df.

A case fatality rate wrapper function to be passed to the gtsummary::add_stat function, which returns a data frame with multiple rows to be used when location is set to "level" for multi-level categorical data. Calls epikit::case_fatality_rate_df.

An attack rate wrapper function to be passed to the gtsummary::add_stat function, which returns a data frame with a single row to be used with dichotomous data or overall data. Calls epikit::attack_rate.

An attack rate wrapper function to be passed to the gtsummary::add_stat function, which returns a data frame with multiple rows to be used when location is set to "level" for multi-level categorical data. Calls epikit::attack_rate.

A mortality rate wrapper function to be passed to the gtsummary::add_stat function, which returns a data frame with multiple rows to be used when location is set to "level" for multi-level categorical data. Calls epikit::attack_rate.

add_mr(
  gts_object,
  deaths_var,
  population = NULL,
  multiplier = 10^4,
  drop_tblsummary_stat = FALSE
)

add_ar(
  gts_object,
  case_var,
  population = NULL,
  multiplier = 10^4,
  drop_tblsummary_stat = FALSE
)

add_cfr(gts_object, deaths_var)

gt_remove_stat(gts_object, col_name = "stat_0")

add_crosstabs(
  data,
  exposure,
  outcome,
  case_reference = "outcome",
  var_name = NULL,
  show_overall = TRUE,
  exposure_label = NULL,
  outcome_label = NULL,
  var_label = NULL,
  two_by_two = FALSE,
  gt_statistic = "{n}",
  show_N_header = FALSE
)

gt_mh_odds(
  data,
  exposure,
  outcome,
  strata,
  exposure_label = NULL,
  outcome_label = NULL,
  strata_label = NULL
)

add_gt_cfr_stat_label(data, variable, by, deaths_var, ...)

add_gt_cfr_stat_level(data, variable, by, deaths_var, ...)

add_gt_attack_rate_stat_label(
  data,
  variable,
  by = NULL,
  case_var,
  population = population,
  multiplier = 10^4,
  drop_total = TRUE,
  drop_cases = TRUE,
  ...
)

add_gt_attack_rate_level(
  data,
  variable,
  by = NULL,
  case_var,
  population = NULL,
  multiplier = 10^4,
  drop_total = TRUE,
  drop_cases = TRUE,
  ...
)

add_gt_mortality_rate_level(
  data,
  variable,
  by = NULL,
  deaths_var,
  population = NULL,
  multiplier = 10^4,
  drop_total = TRUE,
  drop_deaths = TRUE,
  ...
)

Arguments

gts_object

A data frame, passed by the gtsummary::add_stat function

deaths_var

the name of a logical column in the data that indicates that the case died, is passed as the first argument to epikit::case_fatality_rate_df

multiplier

The base by which to multiply the output:

col_name

the column name from the gtsummary object's table_body to remove

data

A data frame, passed by the gtsummary::add_stat function.

exposure

column name to use as the exposure variable, must be logical class

outcome

column name to use as the outcome variable, must be logical class

show_overall

Logical argument to include overall column in gtsummary output; defaults to TRUE

exposure_label

label for exposure variable

outcome_label

label for outcome variable

variable

Name of a variable as the outcome of interest, passed by the gtsummary::add_stat function (e.g. illness)

by

Name of a variable for stratifying, passed by the gtsummary::add_stat function (e.g. illness).

...

additional params that may be passed from gtsummary functions.

drop_total

whether or not to include the population column; default TRUE

Value

a single-row gtsummary object with attack rate results cases, population, attack rate, and 95% confidence interval.

a single-row gtsummary object with attack rate results cases, population, attack rate, and 95% confidence interval.

a single row gtsummary object with case fatality rate results for deaths, cases, cfr, and 95% confidence interval.

a gtsummary object without the named column

gtsummary object with case and control counts tabulated by exposure, along with a crude overall odds ratio and odds using the Cochran-Mantel-Haenszel test with 95% confidence interval (https://cran.r-project.org/web/packages/samplesizeCMH/vignettes/samplesizeCMH-introduction.html)

a single row gtsummary object with case fatality rate results for deaths, cases, cfr, and 95% confidence interval.

a single row or multiple row gtsummary object with with case fatality rate results for deaths, cases, cfr, and 95% confidence interval.

a single row gtsummary object with with attack rate results with cases, attack rate, and 95% confidence interval.

a single-row gtsummary object with attack rate results cases, population, attack rate, and 95% confidence interval.

a single-row gtsummary object with attack rate results cases, population, attack rate, and 95% confidence interval.